requirements.txt
flask==2.3.3
firebase-admin==6.5.0
flask-cors==4.0.1
python-dotenv==1.0.1
